rename hosts directory to team and update component names
